pyproject.toml
[project]
name = "lifely"
version = "0.1.0"
description = "Google Calendar 2025 Wrapped - Analyze your year in events, friends, and places"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.0",
    "google-auth-httplib2>=0.1",
    "google-auth-oauthlib>=1.0",
    "python-dateutil>=2.8",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1",
    "pytest>=7.0",
    "mypy>=1.0",
]

[project.scripts]
lifely = "lifely.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lifely"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
